requirements.txt
Django>=2
Pillow 
psycopg2
psycopg2-binary
python-decouple
djangorestframework
pygments
